Moved some things back. Still working on fixing Javascript issues.
